registro componentizado
